Analytics/pyproject.toml
[project]
name = "analytics"
version = "0.1.0"
description = ""
requires-python = ">=3.11"

dependencies = [
    "prometheus-client>=0.23.1",
    "psycopg2-binary",
]
